+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [Unreleased]
 ----------------------
 ##### Removed
-- IE11 build tasks and loading tricks.
+- Example IE11 build tasks and loading tricks.
 
 ##### Changed
 - Improved performance of autocomplete questions inside repeats that use static option lists.